+++ pysys-examples/internal/testcases/PySys_internal_001/Input/NestedTest/run.py
@@ -42,6 +42,14 @@ class PySysTest(BaseTest):
 		self.assertThat('testNoneProperty == "Hello"', testNoneProperty=self.testNoneProperty)
 		self.assertThat('testStringUserData == expected', testStringUserData=self.testStringUserData, expected='Hello ${non-existent}')
 
+		# check coersion based on default value supplied
+		self.assertThat('projectbool is True', projectbool__eval="self.project.getProperty('projectbool', False)")
+		self.assertThat('projectint == 1234', projectint__eval="self.project.getProperty('projectint', -1)")
+		self.assertThat('projectfloat == 456.78', projectfloat__eval="self.project.getProperty('projectfloat', 0.0)")
+		self.assertThat('user_lastname == "Smith"', user_lastname__eval="self.project.getProperty('user_lastname', 'xxx')")
+
+		#self.assertThat('throws1', testStringProperty__eval="self.project.getProperty('projectfloat', -1)")
+		#self.assertThat('throws2', testStringProperty__eval="self.project.getProperty('projectfloat', None)")
 
 	def validate(self):
 		pass